# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Senac Bike Store API - By Condé",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        },
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Retorna a Lista de Clientes.",
                "description": "Retorna o JSON da Lista de Clientes.",
                "operationId": "getClientesList",
                "parameters": [
                    {
                        "name": "ordenacao",
                        "in": "query",
                        "description": "Ordernação dos dados",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientesResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientesResponseException500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Armazena os dados de um novo Cliente.",
                "description": "Armazena e retorna o JSON de um novo Cliente.",
                "operationId": "postCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteNewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException500"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Retorna os dados de um Cliente.",
                "description": "Retorna o JSON de um Cliente.",
                "operationId": "getCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Apaga os dados de um cliente.",
                "description": "Apaga e retorna o JSON de confirmação.",
                "operationId": "deleteCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualiza os dados de um cliente.",
                "description": "Atualiza e retorna o JSON de um Cliente.",
                "operationId": "patchCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClienteNewRequest": {
                "title": "ClienteNewRequest",
                "description": "Requisição enviada via Body para uma inserção ou atualização de um novo Cliente",
                "required": [
                    "email"
                ],
                "properties": {
                    "nome": {
                        "title": "nome",
                        "description": "Nome do Cliente.",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email do Cliente.",
                        "type": "string"
                    },
                    "telefone": {
                        "title": "telefone",
                        "description": "Telefone do Cliente.",
                        "type": "string"
                    },
                    "rua": {
                        "title": "rua",
                        "description": "Rua do endereço do Cliente.",
                        "type": "string"
                    },
                    "cidade": {
                        "title": "cidade",
                        "description": "Cidade do endereço do Cliente.",
                        "type": "string"
                    },
                    "estado": {
                        "title": "estado",
                        "description": "Estado do endereço do Cliente.",
                        "type": "string"
                    },
                    "cep": {
                        "title": "cep",
                        "description": "Cep do endereço do Cliente.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ClienteResponse": {
                "title": "ClienteResponse",
                "description": "Resposta enviada via JSON para os dados de um Cliente",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Cliente retornado."
                    },
                    "cliente": {
                        "$ref": "#/components/schemas/Cliente"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ClienteResponse"
                }
            },
...
            "ClientesResponse": {
                "title": "ClientesResponse",
                "description": "Resposta enviada via JSON para Lista de Clientes",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Lista de clientes retornada."
                    },
                    "clientes": {
                        "title": "clientes",
                        "description": "Lista de Clientes retornada.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Cliente"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ClientesResponse"
                }
            },
            "ClientesResponseException500": {
                "title": "ClientesResponseException500",
                "description": "Resposta enviada via JSON para Lista de Clientes",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 500
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Erro Interno."
                    },
                    "clientes": {
                        "title": "categorias",
                        "description": "Lista de Clientes retornada.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ClientesResponseException500"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}